+++ docs/release-notes.md
@@ -22,6 +22,7 @@ Minor changes:
 Internal changes:
 - rdcore: Rename `--kargs` to `--append-karg` in `zipl` subcommand and support passing it multiple times
 - Only open BLS configs in write mode when modifying
+- s390x: Don't fail and skip initrd update when there is no LUKS2 keys and config
 
 Packaging changes: